+++ nomos-services/network/Cargo.toml
@@ -9,7 +9,6 @@ workspace = true
 
 [dependencies]
 async-trait      = "0.1"
-bytes            = { workspace = true }
 chrono           = { version = "0.4", optional = true }
 cryptarchia-sync = { workspace = true }
 futures          = "0.3"
